requirements.txt
appdirs==1.4.4
asgiref==3.3.4
black==20.8b1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
coverage==5.5
Django==3.2
django-annoying==0.10.6
django-mathfilters==1.0.0
idna==2.10
mypy-extensions==0.4.3
pathspec==0.8.1
Pillow==8.2.0
pytz==2021.1
regex==2021.4.4
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.7.4.3
urllib3==1.26.4
